Refactor cache invalidation logic to improve modularity
Moved cache invalidation functions to a dedicated module to simplify dependencies and avoid circular imports. Updated related imports and implemented pattern-based invalidation in BaseCacheManager to enhance flexibility.
